[ADF-3323] Moved source info classes
