+++ tools/doc/tools/tsInfo.js
@@ -9,15 +9,13 @@ var remark = require("remark");
 var ejs = require("ejs");
 var typedoc_1 = require("typedoc");
 var mdNav_1 = require("../mdNav");
+var ngHelpers_1 = require("../ngHelpers");
 var libFolders = ["core", "content-services", "process-services", "insights"];
 var templateFolder = path.resolve("tools", "doc", "templates");
 var excludePatterns = [
     "**/*.spec.ts"
 ];
 var nameExceptions;
-var undocMethodNames = {
-    "ngOnChanges": 1
-};
 function processDocs(mdCache, aggData, _errorMessages) {
     //initPhase(aggData);
     nameExceptions = aggData.config.typeNameExceptions;
@@ -39,164 +37,6 @@ function showErrors(filename, errorMessages) {
     });
     console.log("");
 }
-var PropInfo = /** @class */ (function () {
-    function PropInfo(rawProp) {
-        var _this = this;
-        this.errorMessages = [];
-        this.name = rawProp.name;
-        this.docText = rawProp.comment ? rawProp.comment.shortText : "";
-        this.docText = this.docText.replace(/[\n\r]+/g, " ").trim();
-        this.defaultValue = rawProp.defaultValue || "";
-        this.defaultValue = this.defaultValue.replace(/\|/, "\\|");
-        this.type = rawProp.type ? rawProp.type.toString().replace(/\s/g, "") : "";
-        this.type = this.type.replace(/\|/, "\\|");
-        this.isDeprecated = rawProp.comment && rawProp.comment.hasTag("deprecated");
-        if (this.isDeprecated) {
-            this.docText = "(**Deprecated:** " + rawProp.comment.getTag("deprecated").text.replace(/[\n\r]+/g, " ").trim() + ") " + this.docText;
-        }
-        if (rawProp.decorators) {
-            rawProp.decorators.forEach(function (dec) {
-                //console.log(dec);
-                if (dec.name === "Input") {
-                    _this.isInput = true;
-                    if (dec.arguments) {
-                        var bindingName = dec.arguments["bindingPropertyName"];
-                        if (bindingName && (bindingName !== ""))
-                            _this.name = bindingName.replace(/['"]/g, "");
-                    }
-                    if (!_this.docText && !_this.isDeprecated) {
-                        _this.errorMessages.push("Warning: Input \"" + rawProp.name + "\" has no doc text.");
-                    }
-                }
-                if (dec.name === "Output") {
-                    _this.isOutput = true;
-                    if (!_this.docText && !_this.isDeprecated) {
-                        _this.errorMessages.push("Warning: Output \"" + rawProp.name + "\" has no doc text.");
-                    }
-                }
-            });
-        }
-    }
-    Object.defineProperty(PropInfo.prototype, "errors", {
-        get: function () {
-            return this.errorMessages;
-        },
-        enumerable: true,
-        configurable: true
-    });
-    return PropInfo;
-}());
-;
-var ParamInfo = /** @class */ (function () {
-    function ParamInfo(rawParam) {
-        this.name = rawParam.name;
-        this.type = rawParam.type.toString().replace(/\s/g, "");
-        this.defaultValue = rawParam.defaultValue;
-        this.docText = rawParam.comment ? rawParam.comment.text : "";
-        this.docText = this.docText.replace(/[\n\r]+/g, " ").trim();
-        this.isOptional = rawParam.flags.isOptional;
-        this.combined = this.name;
-        if (this.isOptional)
-            this.combined += "?";
-        this.combined += ": `" + this.type + "`";
-        if (this.defaultValue !== "")
-            this.combined += " = `" + this.defaultValue + "`";
-    }
-    return ParamInfo;
-}());
-var MethodSigInfo = /** @class */ (function () {
-    function MethodSigInfo(rawSig) {
-        var _this = this;
-        this.errorMessages = [];
-        this.name = rawSig.name;
-        this.returnType = rawSig.type ? rawSig.type.toString().replace(/\s/g, "") : "";
-        this.returnsSomething = this.returnType != "void";
-        if (rawSig.hasComment()) {
-            this.docText = rawSig.comment.shortText + rawSig.comment.text;
-            this.docText = this.docText.replace(/[\n\r]+/g, " ").trim();
-            if (!this.docText) {
-                this.errorMessages.push("Warning: method \"" + rawSig.name + "\" has no doc text.");
-            }
-            this.returnDocText = rawSig.comment.returns;
-            this.returnDocText = this.returnDocText ? this.returnDocText.replace(/[\n\r]+/g, " ").trim() : "";
-            if (this.returnDocText.toLowerCase() === "nothing") {
-                this.returnsSomething = false;
-            }
-            if (this.returnsSomething && !this.returnDocText) {
-                this.errorMessages.push("Warning: Return value of method \"" + rawSig.name + "\" has no doc text.");
-            }
-            this.isDeprecated = rawSig.comment.hasTag("deprecated");
-        }
-        this.params = [];
-        var paramStrings = [];
-        if (rawSig.parameters) {
-            rawSig.parameters.forEach(function (rawParam) {
-                if (!rawParam.comment || !rawParam.comment.text) {
-                    _this.errorMessages.push("Warning: parameter \"" + rawParam.name + "\" of method \"" + rawSig.name + "\" has no doc text.");
-                }
-                var param = new ParamInfo(rawParam);
-                _this.params.push(param);
-                paramStrings.push(param.combined);
-            });
-        }
-        this.signature = "(" + paramStrings.join(", ") + ")";
-    }
-    Object.defineProperty(MethodSigInfo.prototype, "errors", {
-        get: function () {
-            return this.errorMessages;
-        },
-        enumerable: true,
-        configurable: true
-    });
-    return MethodSigInfo;
-}());
-var ComponentInfo = /** @class */ (function () {
-    function ComponentInfo(classRef) {
-        var _this = this;
-        var props = classRef.getChildrenByKind(typedoc_1.ReflectionKind.Property);
-        var accessors = classRef.getChildrenByKind(typedoc_1.ReflectionKind.Accessor);
-        this.properties = props.concat(accessors).map(function (item) {
-            return new PropInfo(item);
-        });
-        var methods = classRef.getChildrenByKind(typedoc_1.ReflectionKind.Method);
-        this.methods = [];
-        methods.forEach(function (method) {
-            if (!(method.flags.isPrivate || method.flags.isProtected || undocMethodNames[method.name])) {
-                method.signatures.forEach(function (sig) {
-                    _this.methods.push(new MethodSigInfo(sig));
-                });
-            }
-        });
-        this.hasInputs = false;
-        this.hasOutputs = false;
-        this.properties.forEach(function (prop) {
-            if (prop.isInput)
-                _this.hasInputs = true;
-            if (prop.isOutput)
-                _this.hasOutputs = true;
-        });
-        this.hasMethods = methods.length > 0;
-    }
-    Object.defineProperty(ComponentInfo.prototype, "errors", {
-        get: function () {
-            var combinedErrors = [];
-            this.methods.forEach(function (method) {
-                method.errors.forEach(function (err) {
-                    combinedErrors.push(err);
-                });
-            });
-            this.properties.forEach(function (prop) {
-                prop.errors.forEach(function (err) {
-                    combinedErrors.push(err);
-                });
-            });
-            return combinedErrors;
-        },
-        enumerable: true,
-        configurable: true
-    });
-    return ComponentInfo;
-}());
 function initPhase(aggData) {
     nameExceptions = aggData.config.typeNameExceptions;
     var app = new typedoc_1.Application({
@@ -211,15 +51,21 @@ function initPhase(aggData) {
     aggData.projData = app.convert(sources);
 }
 function updateFile(tree, pathname, aggData, errorMessages) {
-    var compName = angNameToClassName(path.basename(pathname, ".md"));
-    var classRef = aggData.projData.findReflectionByName(compName);
+    /*
+    let compName = angNameToClassName(path.basename(pathname, ".md"));
+    let classRef = aggData.projData.findReflectionByName(compName);
+
     if (!classRef) {
         // A doc file with no corresponding class (eg, Document Library Model).
         return false;
     }
-    var compData = new ComponentInfo(classRef);
-    var classTypeMatch = compName.match(/component|directive|service/i);
+
+    let compData = new ComponentInfo(classRef);
+    */
+    var className = ngHelpers_1.ngNameToClassName(path.basename(pathname, ".md"));
+    var classTypeMatch = className.match(/component|directive|service/i);
     if (classTypeMatch) {
+        var compData = aggData.classInfo[className];
         var classType = classTypeMatch[0].toLowerCase();
         // Copy docs back from the .md file when the JSDocs are empty.
         var inputMD = getPropDocsFromMD(tree, "Properties", 3);
@@ -249,23 +95,33 @@ function updateFile(tree, pathname, aggData, errorMessages) {
 function initialCap(str) {
     return str[0].toUpperCase() + str.substr(1);
 }
-function angNameToClassName(rawName) {
+/*
+function angNameToClassName(rawName: string) {
     if (nameExceptions[rawName])
         return nameExceptions[rawName];
+
     var name = rawName.replace(/\]|\(|\)/g, '');
+    
     var fileNameSections = name.split('.');
     var compNameSections = fileNameSections[0].split('-');
+    
     var outCompName = '';
+    
     for (var i = 0; i < compNameSections.length; i++) {
         outCompName = outCompName + initialCap(compNameSections[i]);
     }
+    
     var itemTypeIndicator = '';
+    
     if (fileNameSections.length > 1) {
         itemTypeIndicator = initialCap(fileNameSections[1]);
     }
+    
     var finalName = outCompName + itemTypeIndicator;
+    
     return finalName;
 }
+*/
 function getPropDocsFromMD(tree, sectionHeading, docsColumn) {
     var result = {};
     var nav = new mdNav_1.MDNav(tree);
